Moved GetJobException and SubmitJobException to arrebol package
