+++ src/main/java/org/fogbowcloud/saps/engine/core/scheduler/executor/arrebol/request/ArrebolRequestsHelper.java
@@ -10,8 +10,8 @@ import org.apache.http.entity.StringEntity;
 import org.apache.log4j.Logger;
 import org.fogbowcloud.saps.engine.core.scheduler.executor.arrebol.dtos.JobRequestDTO;
 import org.fogbowcloud.saps.engine.core.scheduler.executor.arrebol.dtos.JobResponseDTO;
-import org.fogbowcloud.saps.engine.core.scheduler.arrebol.exceptions.GetJobException;
-import org.fogbowcloud.saps.engine.core.scheduler.arrebol.exceptions.SubmitJobException;
+import org.fogbowcloud.saps.engine.core.scheduler.executor.arrebol.exceptions.GetJobException;
+import org.fogbowcloud.saps.engine.core.scheduler.executor.arrebol.exceptions.SubmitJobException;
 import org.fogbowcloud.saps.engine.core.scheduler.executor.arrebol.exceptions.ArrebolConnectException;
 import org.fogbowcloud.saps.engine.core.scheduler.executor.arrebol.models.ArrebolQueue;
 import org.fogbowcloud.saps.engine.core.scheduler.executor.arrebol.http.HttpWrapper;
